trgil_6/requirements.txt
numpy == 1.20.2
Pillow == 8.2.0
scikit-image ==	0.18.1
PIL-Tools ==1.1.0
imageio == 2.9.0
